Review logic and comments
