+++ packages/validator/src/util/index.ts
@@ -1,4 +1,5 @@
 export * from "./aggregator";
+export * from "./clock";
 export * from "./error";
 export * from "./httpClient";
 export * from "./url";
